+++ backend/canisters/community/impl/src/timer_job_types.rs
@@ -8,8 +8,6 @@ use chat_events::MessageContentInternal;
 use constants::{DAY_IN_MS, MINUTE_IN_MS, NANOS_PER_MILLISECOND, SECOND_IN_MS};
 use group_chat_core::AddResult;
 use ledger_utils::process_transaction;
-use rand::rngs::StdRng;
-use rand::SeedableRng;
 use serde::{Deserialize, Serialize};
 use tracing::{error, info};
 use types::{
@@ -32,7 +30,6 @@ pub enum TimerJob {
     MarkP2PSwapExpired(MarkP2PSwapExpiredJob),
     MarkVideoCallEnded(MarkVideoCallEndedJob),
     JoinMembersToPublicChannel(JoinMembersToPublicChannelJob),
-    DedupeMessageIds(DedupeMessageIdsJob),
 }
 
 #[derive(Serialize, Deserialize, Clone)]
@@ -149,11 +146,6 @@ pub struct JoinMembersToPublicChannelJob {
     pub members: Vec<UserId>,
 }
 
-#[derive(Serialize, Deserialize, Clone, Default)]
-pub struct DedupeMessageIdsJob {
-    iteration: u32,
-}
-
 impl Job for TimerJob {
     fn execute(self) {
         if can_borrow_state() {
@@ -175,7 +167,6 @@ impl Job for TimerJob {
             TimerJob::MarkP2PSwapExpired(job) => job.execute(),
             TimerJob::MarkVideoCallEnded(job) => job.execute(),
             TimerJob::JoinMembersToPublicChannel(job) => job.execute(),
-            TimerJob::DedupeMessageIds(job) => job.execute(),
         }
     }
 }
@@ -520,34 +511,5 @@ impl JoinMembersToPublicChannelJob {
                     .enqueue_job(TimerJob::JoinMembersToPublicChannel(self), now, now);
             }
         }
-    }
-}
-
-impl Job for DedupeMessageIdsJob {
-    fn execute(mut self) {
-        mutate_state(|state| {
-            let mut complete = true;
-            let seed = state.env.entropy();
-            let mut rng = StdRng::from_seed(seed);
-            for channel in state.data.channels.iter_mut() {
-                match channel.chat.events.fix_duplicate_message_ids(&mut rng, None) {
-                    Some(true) => {}
-                    Some(false) => {
-                        complete = false;
-                        break;
-                    }
-                    None => error!("Failed to dedupe messageIds"),
-                }
-            }
-            if complete {
-                state.data.message_ids_deduped = true;
-            } else if self.iteration < 100 {
-                self.iteration += 1;
-                let now = state.env.now();
-                state.data.timer_jobs.enqueue_job(TimerJob::DedupeMessageIds(self), now, now);
-            } else {
-                error!("Failed to dedupe messageIds after 100 iterations");
-            }
-        })
     }
 }