+++ backend/canisters/group_index/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Fixed
+
+- Avoid retrying c2c call if recipient canister is uninstalled ([#7302](https://github.com/open-chat-labs/open-chat/pull/7302))
+
+## [[2.0.1584](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1584-group_index)] - 2025-01-24
+
 ### Added
 
 - Add facility to set/revoke community/group verification via proposal ([#7240](https://github.com/open-chat-labs/open-chat/pull/7240))